BackEnd/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
lxml==6.0.2
multidict==6.7.1
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-docx==1.2.0
requests==2.32.5
sqlparse==0.5.3
twilio==9.10.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
yarl==1.23.0
